move eager-prebid-check
