+++ bundle/src/projects/commercial/modules/comments-expanded-advert.ts
@@ -1,7 +1,7 @@
 import { adSizes, createAdSlot } from '@guardian/commercial-core';
 import { AD_LABEL_HEIGHT } from '@guardian/commercial-core/dist/cjs/constants';
 import { log } from '@guardian/libs';
-import { isInEagerPrebidVariant } from 'common/modules/experiments/tests/eager-prebid-check';
+import { isInEagerPrebidVariant } from 'common/modules/experiments/eager-prebid-check';
 import fastdom from '../../../lib/fastdom-promise';
 import { commercialFeatures } from '../../common/modules/commercial/commercial-features';
 import { addSlot } from './dfp/add-slot';
