+++ bundle/src/projects/commercial/modules/dfp/fill-advert-slots.ts
@@ -2,8 +2,8 @@ import type { SizeMapping } from '@guardian/commercial-core';
 import { adSizes, createAdSize } from '@guardian/commercial-core';
 import { log } from '@guardian/libs';
 import { isInVariantSynchronous } from 'common/modules/experiments/ab';
+import { isInEagerPrebidVariant } from 'common/modules/experiments/eager-prebid-check';
 import { billboardsInMerch } from 'common/modules/experiments/tests/billboards-in-merch';
-import { isInEagerPrebidVariant } from 'common/modules/experiments/tests/eager-prebid-check';
 import { getCurrentBreakpoint } from 'lib/detect-breakpoint';
 import { commercialFeatures } from '../../../common/modules/commercial/commercial-features';
 import { requestBidsForAds } from '../header-bidding/request-bids';
