+++ bundle/src/projects/commercial/modules/liveblog-adverts.ts
@@ -1,6 +1,6 @@
 import { adSizes, createAdSlot } from '@guardian/commercial-core';
 import { log } from '@guardian/libs';
-import { isInEagerPrebidVariant } from 'common/modules/experiments/tests/eager-prebid-check';
+import { isInEagerPrebidVariant } from 'common/modules/experiments/eager-prebid-check';
 import { getCurrentBreakpoint } from 'lib/detect-breakpoint';
 import { getUrlVars } from 'lib/url';
 import fastdom from '../../../lib/fastdom-promise';
