+++ bundle/src/projects/common/modules/experiments/eager-prebid-check.ts
@@ -1,6 +1,10 @@
 import { memoize } from 'lodash-es';
-import { getSynchronousTestsToRun } from '../ab';
-import { eagerPrebid } from './eager-prebid';
+import { getSynchronousTestsToRun } from './ab';
+import { eagerPrebid } from './tests/eager-prebid';
+
+/**
+ * These functions can't be in the eager-prebid.ts file because there is a circular dependency
+ */
 
 export const getEagerPrebidVariant = memoize((): string => {
 	const tests = getSynchronousTestsToRun();
